Updated Factorial Recursion function

+++ J_Factorial.cpp
@@ -1,14 +1,14 @@
 #include<bits/stdc++.h>
 using namespace std;
 
-int fun(int n){
+long long fun(long long n){
     if(n == 0 || n == 1) return 1;
     return n* fun(n-1);
 }
 
 int main(){
-    int n;
+    long long n;
     cin >> n;
-    int res = fun(n);
+    long long res = fun(n);
     cout << res << endl;
 }
